+++ src/app/(home)/page.tsx
@@ -1,7 +1,7 @@
 import ArenaDescription from "@/components/arenaDescription";
 import ArenaMenu from "@/components/arenaMenu";
 import ArenaChat from "@/containers/arenaChat";
-import { ArenaMenuKey } from "@/type";
+import { ArenaMenuKey } from "@/types/type";
 import MaintenanceMsg from "@/components/maintenance";
 
 export default function Home() {
